pyproject.toml
[tool.coverage.run]
plugins = ['django_coverage_plugin']
omit = ["manage.py", "tasks/*"]

[tool.poetry]
name = "role-play"
version = "0.10.0"
description = "Role playing game."
authors = ["Eric Grabarczyk <eric.grabarczyk@gmail.com>"]
license = "GNU Affero General Public License v3.0"
readme = "README.md"
packages = [{ include = "role_play" }]

[tool.poetry.dependencies]
python = "^3.11"
async-timeout = "*"
celery = "*"
channels = "*"
channels-redis = "*"
daphne = "*"
dice = "*"
dj-database-url = "*"
django = "~=5.1"
django-celery-beat = "*"
django-extensions = "*"
django-model-utils = "*"
django-viewflow = "*"
email-validator = "*"
gunicorn = "*"
invoke = "*"
psycopg = { extras = ["binary"], version = "*" }
pydantic = "*"
pyyaml = "*"
redis = "*"
whitenoise = "*"

[tool.poetry.group.dev.dependencies]
coverage = "*"
django-coverage-plugin = "*"
django-debug-toolbar = "*"
factory_boy = "*"
faker = "*"
freezegun = "*"
pre-commit = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-django = "*"

[tool.ruff]
exclude = ["**/role_play/asgi.py", "**/role_play/urls.py", "**/role_play/settings/*"]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
